refactor(registration-react): Rename items-list to dataset-items-list
