+++ applications/registration-react/src/pages/dataset-list-page/items-list/dataset-items-list.scss
